+++ src/SystemsUnderTest/Sut.WinForms.WorkflowsTest/ScreenObjects/AddressWizardPage.cs
@@ -8,27 +8,27 @@ namespace Sut.WinForms.WorkflowsTest.ScreenObjects
     {
         public string Address
         {
-            set { Find<WinEdit>(By.ControlName("textBoxAddress").AndName("Address")).Text = value; }
+            set { Find<WinWindow>(By.ControlName("textBoxAddress")).Find<WinEdit>(By.Name("Address")).Text = value; }
         }
 
         public string City
         {
-            set { Find<WinEdit>(By.ControlName("textBoxCity").AndName("City")).Text = value; }
+            set { Find<WinWindow>(By.ControlName("textBoxCity")).Find<WinEdit>(By.Name("City")).Text = value; }
         }
 
         public string PostalCode
         {
-            set { Find<WinEdit>(By.ControlName("textBoxPostalCode").AndName("PostalCode")).Text = value; }
+            set { Find<WinWindow>(By.ControlName("textBoxPostalCode")).Find<WinEdit>(By.Name("PostalCode")).Text = value; }
         }
 
         public string State
         {
-            set { Find<WinEdit>(By.ControlName("textBoxState").AndName("State")).Text = value; }
+            set { Find<WinWindow>(By.ControlName("textBoxState")).Find<WinEdit>(By.Name("State")).Text = value; }
         }
 
         public FinishedWizardPage ClickNext()
         {
-            Find<WinButton>(By.Name("Next")).Click();
+            Find<WinWindow>(By.ControlName("buttonNext")).Find<WinButton>(By.Name("Next")).Click();
             return GetScreenObject<FinishedWizardPage>();
         }
     }